# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Aerial Housing Detection API
  description: API para detecção de habitações em imagens aéreas com cruzamento CadUnico
  version: 1.0.0
  contact:
    name: Prefeitura de Paulista-PE
    email: ti@paulista.pe.gov.br
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8000
    description: Desenvolvimento local
  - url: https://staging.aerial.paulista.pe.gov.br
    description: Staging
  - url: https://aerial.paulista.pe.gov.br
    description: Producao
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/v1/monitoring/alerts:
    get:
      summary: Alertas ativos
      operationId: getAlerts
      tags: [monitoring]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, acknowledged, resolved, all]
            default: active
      responses:
        "200":
          description: Lista de alertas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /api/v1/monitoring/alerts/{alert_id}/acknowledge:
    post:
      summary: Reconhecer alerta
      operationId: acknowledgeAlert
      tags: [monitoring]
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alerta reconhecido
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
...
    Alert:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        severity:
          type: string
          enum: [warning, critical]
        message:
          type: string
        status:
          type: string
          enum: [active, acknowledged, resolved]
        created_at:
          type: string
          format: date-time
...